Learn To Cook Xbox Edition
Due to recent world events I have had to move towards a new version of LTC and this is LTC Xbox Edititon. I will be making a prototype to further release down the line however due to time restraints this will be pc based.

Getting my head around the new Unity System has taken time, this is why the push is so late. However in that time I have been able to also rig up a  pause menu. It is slow progress however alas it is still moving forward