Major Style
Adding all of the materials for the items in the game. Now as they all have a uniform look the game looks more polished.

Next Changes will be to implement the book system inputted to the console version and possibly changing how the game works in regards to adding items to unlock recipes all in one scene