Major Changes - Combination System Works
The main system for LTC VR combining the different ingredients on the placemats now works. This is achieved through the use of lists and checking the tags. This also happens to be modular so it is easy to plug in different recipes. The next idea is to make sure the ingredients in the game are extensive to make recipes with a few dropped in there for humour.

Really happy with current progress :)