Functions Working
Managing to get the game working so many different combinations of items can be added. Then when the submit button is pressed (and being correct) the items will be deleted after a short pause. This next will be tied in with a recipe gallery (likely to be a poster system). Overall a good 70 % through on development of the game itself