Informative Process
Over the last couple days I have been adding in the posters to guide the user in what to do in the game. This takes time due to the fact that they have to have their own descriptions. In addition to this all of the functions within the game now work so it is a case of sorting the art direction as well as adding more audio. There is a sound that plays when the answer is correct so one needs to be set up for it. Overall progress is going well!